debug plate & pump class, edit led functions (mainly rainbow)

+++ src/Pump.cpp
@@ -8,10 +8,10 @@
 
 
 //750ml준다고 하고 1초에 물 10ml나온다 가정
-inline double Pump::cal_pump(double amount) 
+inline int Pump::cal_pump(int amount) 
 {
 	const int amount_per_sec_ratio = 10;
-	double pump_time = amount / amount_per_sec_ratio; //양->초
+	int pump_time = amount / amount_per_sec_ratio; //양->초
 	return pump_time;
 }
 
@@ -32,7 +32,7 @@ void Pump::stop_pump()
 }
 // 왜 같지?? - 모터드라이브 스펙시트 보고 바꿔야 할 듯..!
 
-void Pump::work_pump(double amount) {
+void Pump::work_pump(int amount) {
 	start_pump();
 	delay(cal_pump(amount));//변환 시간 입력
 	stop_pump();
